info_quest_backend/requirements.txt
fastapi>=0.77.1
pydantic>=1.0.0
PyYAML>=6.0